// src/openapi/model/sTOCKTYPE.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type STOCKTYPE = 'packageMaterial' | 'defectiveProduct' | 'tools' | 'inactiveStock' | 'gauge' | 'material' | 'finishedProduct' | 'semiFinishedProduct' | 'outsourcing' | 'machine' | 'assemble' | 'mfgBuffer' | 'autoliv' | 'ufo';

export const STOCKTYPE = {
    PackageMaterial: 'packageMaterial' as STOCKTYPE,
    DefectiveProduct: 'defectiveProduct' as STOCKTYPE,
    Tools: 'tools' as STOCKTYPE,
    InactiveStock: 'inactiveStock' as STOCKTYPE,
    Gauge: 'gauge' as STOCKTYPE,
    Material: 'material' as STOCKTYPE,
    FinishedProduct: 'finishedProduct' as STOCKTYPE,
    SemiFinishedProduct: 'semiFinishedProduct' as STOCKTYPE,
    Outsourcing: 'outsourcing' as STOCKTYPE,
    Machine: 'machine' as STOCKTYPE,
    Assemble: 'assemble' as STOCKTYPE,
    MfgBuffer: 'mfgBuffer' as STOCKTYPE,
    Autoliv: 'autoliv' as STOCKTYPE,
    Ufo: 'ufo' as STOCKTYPE
};